AI-based screening of pulomary tuberculosis
Screening of the pulmonary tuberculosis based on the chest radiographs. The model is trained in such a way that it takes the input as a chest radiograph and gives the output whether it is affected from tuberculosis or not.The model is trained using the dataset available from the internet. It consists of  over 1300 chest radiographs combined of affected or unaffected from montgomery and  shenzen hospital.